requirements.txt
nonebot2[fastapi]>=2.4.3
nonebot-adapter-onebot>=2.4.6
httpx>=0.27.0
pyodbc>=4.0.30